+++ Plugins/InAppNotifications/CHANGELOG.md
@@ -1,4 +1,8 @@
 # [InAppNotifications](https://1lighty.github.io/BetterDiscordStuff/?plugin=InAppNotifications "InAppNotifications") Changelog
+### 1.0.1
+- Emotes and custom emotes now show properly (not BD emotes tho lol)
+- Changed title to show channel name and guild name instead of category
+
 ### 1.0.0
 - In-app notifications will show when someone sends a message!
 - It's tied to your notification settings of Discord, which can be accessed by right clicking a server, channel, or muting a DM.